requirements.txt
django==4.1
django-extensions==3.1.5
djangorestframework==3.13.1
django-storages
google-api-core
google-auth
google-auth-httplib2
google-auth-oauthlib
google-api-python-client
rollbar
google-cloud-core
google-cloud-storage
requests-mock
google-crc32c
google-resumable-media
googleapis-common-protos
Pillow
requests
liqpay-sdk-python3
selenium==4.8.3
pendulum
python-telegram-bot==13.13
honcho
redis==4.6
asyncpg
psycopg2-binary
dj-database-url
pytest==7.1.3
pytest-django==4.5.2
django-polymorphic
celery==5.3.1
flower
tzdata
folium
shapely
python-gettext
gunicorn
gevent
backports.zoneinfo
pyjwt==2.8.0
cryptography
django_celery_beat
django-debug-toolbar
sentry-sdk[django]
sentry-sdk[celery]
django-cors-headers
